requirements.txt
streamlit
audio-recorder-streamlit
openai
pandas
pillow

